client/requirements.txt
streamlit==1.44
streamlit-aggrid==0.3.4.post3
streamlit-option-menu==0.3.6
boto3==1.38.18
pandas==2.2.3
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-dotenv==1.1.0
nest-asyncio==1.6.0

langchain==0.3.20
langchain-aws==0.2.12
langchain-openai>=0.0.3
langchain-anthropic>=0.1.1
langchain-google-genai>=2.1.2
langchain-mcp-adapters==0.0.7
langchain_groq>=0.3.6
langgraph==0.3.30